+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - Fix rename panel files to reflect current version.
 
+## Added
+- `coarsened_pmds_layout` method for more performant cell layout computation.
+
 ## [0.27.2] - 2026-05-29
 
 ### Fixed